+++ internal/rules/repository_impl_test.go
@@ -18,6 +18,7 @@ package rules
 
 import (
 	"context"
+	"github.com/dadrus/heimdall/internal/x/radixtree"
 	"net/http"
 	"net/url"
 	"testing"
@@ -30,7 +31,6 @@ import (
 
 	"github.com/dadrus/heimdall/internal/heimdall"
 	mocks2 "github.com/dadrus/heimdall/internal/heimdall/mocks"
-	"github.com/dadrus/heimdall/internal/radixtree"
 	"github.com/dadrus/heimdall/internal/rules/event"
 	"github.com/dadrus/heimdall/internal/rules/rule"
 	"github.com/dadrus/heimdall/internal/rules/rule/mocks"
